firstsite/requirements.txt
asgiref==3.5.2
Babel==2.12.1
backports.zoneinfo==0.2.1;python_version<"3.9"
dj-database-url==1.0.0
Django==4.0.8
django-countries==7.4.2
django-filebrowser==4.0.1
django-filter==22.1
django-grappelli==3.0.3
django-heroku==0.3.1
django-jinja==2.10.2
django-jquery==3.1.0
django-phonenumber-field==7.0.0
django-taggit==3.0.0
django-tinymce==3.4.0
djangorestframework==3.14.0
docxcompose==1.4.0
docxtpl==0.16.6
gunicorn==20.1.0
Jinja2==3.1.2
lxml==4.9.2
MarkupSafe==2.1.1
phonenumbers==8.13.0
Pillow==9.3.0
psycopg2==2.9.4
python-docx==0.8.11
pytz==2022.6
pyuca==1.2
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.4.0
tzdata==2022.1
whitenoise==6.2.0